backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
pgvector>=0.3
pydantic>=2.6
pydantic-settings>=2.1
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
langchain>=0.1
langchain-openai>=0.0.8
openai>=1.12
httpx>=0.27
python-multipart>=0.0.9
